Move each OAuth page body into its own file
Follow-up to the shared-chrome extraction: the endpoint handlers were
still carrying ~365 lines of inline HTML/CSS/JS for the four pages
(including a 12-line favicon SVG path and the password-rules eval
script). Moves each page's body + page-specific CSS + scripts into its
own class under Api/Helpers/OAuthPages/:

  OAuthLoginPage.cs        (103 lines — includes inline logo SVG, GitHub button, or-divider CSS)
  OAuthRegisterPage.cs     ( 97 lines — password rules list + eval script + TOS)
  OAuthVerifyEmailPage.cs  ( 80 lines — OTP input + resend controls)
  OAuthConsentPage.cs      ( 84 lines — logo/permissions panel + deny button)

Each page class takes raw values (csrfToken, returnUrl, email, clientId,
clientName, error) and handles its own HTML/URL encoding. Endpoint
handlers shrink to ~5 lines: parse inputs, call Page.Render(), return
Results.Content.

OAuthEndpoints.cs: 1127 -> 819 lines. With the earlier shared-chrome
refactor that's 1341 -> 819 (-522) for the endpoints file — it's now
pure routing + POST business logic, zero HTML.
